+++ src/main/java/instruments/Drums.java
@@ -1,7 +1,7 @@
 package instruments;
 
 public class Drums extends Instrument{
-    public Drums(String manufacturer, String material, int buyPrice, int sellPrice){
+    public Drums(String manufacturer, String material, double buyPrice, double sellPrice){
         super(manufacturer, material, buyPrice, sellPrice);
     }
