+++ src/main/java/instruments/Guitar.java
@@ -1,7 +1,7 @@
 package instruments;
 
 public class Guitar extends Instrument{
-    public Guitar(String manufacturer, String material, int buyPrice, int sellPrice){
+    public Guitar(String manufacturer, String material, double buyPrice, double sellPrice){
         super(manufacturer, material, buyPrice, sellPrice);
     }
